requirements.txt
functions-framework==3.*
flask>=3.0.0
pydantic>=2.7.0
pydantic-settings>=2.3.0
google-genai>=1.0.0
ollama>=0.3.0
google-cloud-aiplatform>=1.60.0
google-cloud-firestore>=2.16.0
google-cloud-storage>=2.16.0
google-cloud-documentai>=2.24.0
Pillow>=10.3.0
PyJWT>=2.8.0
requests>=2.31.0
python-dotenv>=1.0.0
bcrypt>=4.1.0
httpx>=0.27.0
python-multipart>=0.0.9
reportlab>=4.2.0
pdfplumber>=0.11.0
python-docx>=1.1.0
pytest>=8.2.0
pytest-asyncio>=0.23.0
